Added config/sys/data_zone_config/list_all.json
This is a version of config/sys/data_zone_config/list.json, but including
all the zones and not just those that should be drawn on the zones map. This is
to support the Django data API that needs a source for all zone
metadata. See

  https://github.com/ijl20/tfc_web/commit/3c2cc72d8d3150df54cca84db8e24c37847db7e2

This file has an additional boolean field "zone.map", not present in
list.json or the other *.json files, which is accessible via the data API
and which can be used to decide if the zone should be drawn on the zones map.