Enable transfer of ACME HTTP-01 challenges
Rework LetsEncrypt certificate management by creating a USERV service
that can copy a ACME HTTP-01 challenge from a machine that's requesting
a certificate the to (perhaps the same, perhaps different) machine that's
currently supporting that domain. This way, for example, every machine
in the cluster can be provisioned with its own certificate for
smartcambridge.org and be ready to take over the service whenever
traffic is directed to it. Such a certificate will also be correctly
auto-renewed as necessary.

Authorisation to use the copying service, and so to request certificates
for any domains supported by he cluster, is granted by having access to
the SSH private key used to invoke it.

The setup involves a dedicated user (certbot-challlenge), a USERV service,
some custom nginx config to make a single /.well-known/acme-challenge
available from multiple servers, and some scripts to run
`certbot certonly ---manual` with appropriate arguments to request
certificates.

Setup instructions are available in README.md