Rename includes_tfc to includes2
Makes the naming slightly more consistens, with tfc_prod.conf/includes
vs. tfc_prod2.conf/includes2

The logic behind creating a new '.conf' file and include directory is
that the old an new configuration can then coexist without affecting
each other with only a symlink is sites-enabled needing to be changed
to switch between the two. This should make deploying this on in production
easier.