Rework https redirection so decision is in include files
It's unfortunately too broad-brush to unconditionally redirect
all http requests to https because we may need to support http feeds
from elsewhere (and for the time being we need to support http feeds
from other nodes in the cluster because that's how they are configured
and they don't understand redirects).

This edit makes all hosts dual http/https and leaves it up to the
included location blocks to decide if they want the traffic that they
handle redirected to https:. If they do, they can add

        # Redirect to https
        if ($scheme != "https") {
          return 301 https://$host$request_uri;
        }

At the moment only the tfc_web include file is doint this.