Tweak monit config; add tfc_web-server monitoring
Modify the monit config to send email to/from admin@smartcambridge.org,
and to omit mailing on starts, stops and user initiated actions; move
the internal http server config to an included file so the main monitrc
file can go into git.

Add a config fragment to monitor http and https on the tfc servers and
smartcambridge.org.