Restart nginx from deploy hook, not cleanup
The certbot renewal cleanup hook is run after each domain validation,
but probably before the certificate is actually issued. So it's the wrong
place to reload nginx to pick up a new certificate. The right place to do
this is ina  deploy hook, and the easiest way to specify a deploy hook
is by putting a script in /etc/letsencrypt/renewal-hooks/deploy/

Hopefully fixes #21, but only time will tell.