Add explicit handling for requests targeting unknown vhosts
Prior to this edit, requests containing unknown names in their Host:
header were routed to the the server in the tfc_prod2.conf
configuration. This generally results in complaints from Django that the
host was unrecognised.

This edit adds explicit configuration for a server block that will
handle such requests and return an informative error page.

Doing this over https: is harder because nginx requires a key and
certificate, but such a certificate would need to list all possible
'unexpected' host names. The best we can do is used an obviously bogus
self-signed certificate. If someone ignores the resulting browser
warning they will at least get the normal informative error page.