Create missisng ~tfc_prod/tfc_prod directory
The tfc_prod repository used to be checked out to ~tfc_prod/tfc_prod.
We don't do that now, but we put a few things here for the benefit
of other things that expect it.

Fixup the destinations for tfc_prod's secrets.py and run.sh

Copy tfc_prod's tools directory to ~tfc_prod/tfc_prod/tools