Add www.smartcambridge.org and IP address, push keys onto one line
The LetsEncrypt automation needs to be able to ssh as
'www.smartcambridge.org' as well as 'smartcambrifgr.org', so add that to
the generated ssh_known_hosts file.

ssh checks IP address and well as hostname against authorized_keys by
default. While it's not a fault if its missing, it adds it to
~/.ssh/authorized_keys and this will be a pain if our
hostname-to-IP-address mapping changes in the future. So add each
server's IP address to the file too.

At the same time simplify the file by listing all the hostnames and the
IP address against a single copy of each key.