Only monitor the web server here and on smartcambridge.org
Previously all servers monitored the web server on all servers,
which causes a lot of noise (especially during development). This
edit cases each server to only monitor its own server and
smartcmbridge.org.

This requires the tfc_web_servers config to become a template so the
target host name can be targeted. We can't target 'localhost' because
that won't work for https: