Rework monitoring of incoming SIRIVM
* Rename tfc_sirivm config file as tfc_cloudamber, and make it monitor
the last change date on the directory /media/tfc/cloudamber/sirivm/data_monitor.
This checks that we have incomming XML. This should b linked
into conf-enabled only on the machine receiving this data (which will be the
one acting as smartcambridge.org)
* Rework the cronjobs to start/stop the aggressive monitoring in tfc_cloudamber
to take account of the changes in name
* Add a new tfc_sirivm_json config that monitors the last change date on
the /media/tfc/sirivm_json/data_monitor directory. This indirectly
checks that we are generating archived sirivm data in json.