Add nginx config for downoad API files
Add config to serve compressed archive files for the download API
from /media/tfc/download_api under URLs starting
http[s]://...../api/download_files

Access to these files requires that the user is authenticated to
Django and has accepted the SmartCambridge T&Cs. This is enforced
using Nginx's auth_request module which makes a call to a probe
endpoint (implemented in tfc_web as part of commit b9fef6c3).
Users that fail the test are redirected to the existing
/api/login-and-agree endpoint to complete the authentication
process, after which they are returned to the URL they originally
requested.

The archived files are served direct by Nginx to avoid the
overhead of going through Python ad Django for what could be
substantial files.